Moved scene audio logic to new SceneAudio script -Created a SceneAudio prefab and added it to each scene
