+++ Assets/Prefabs/AudioManager.prefab
@@ -1273,14 +1273,6 @@ MonoBehaviour:
   - {fileID: 3313367527656396670}
   - {fileID: 3313367528648828627}
   bgMusic:
-  - {fileID: 3313367527789373859}
-  - {fileID: 3313367527929387678}
-  - {fileID: 3313367528366674734}
-  - {fileID: 3313367527486158810}
-  - {fileID: 3313367528718562564}
-  - {fileID: 3313367528594909108}
-  - {fileID: 3313367527656396670}
-  - {fileID: 3313367528648828627}
   - {fileID: 3313367527804858753}
   - {fileID: 3313367527617200816}
   - {fileID: 3313367528062199895}
