+++ Assets/Scripts/AudioManager.cs
@@ -5,30 +5,21 @@ using UnityEngine;
 public class AudioManager : MonoBehaviour
 {
 
+    // public variables
     public AudioSource[] sfx;
     public AudioSource[] bgMusic;
-
     public static AudioManager instance;
 
-    // Start is called before the first frame update
     void Start()
     {
         instance = this;
         DontDestroyOnLoad(this.gameObject);
     }
 
-    // Update is called once per frame
-    void Update()
-    {
-        // if (Input.GetKeyDown(KeyCode.T)) {
-        //     PlaySFX(0);
-        // }
-    }
-
     public void PlaySFX(int sound)
     {
         if (sound < sfx.Length) {
-            // sfx[sound].Play();
+            sfx[sound].Play();
         }
     }
 
@@ -38,8 +29,6 @@ public class AudioManager : MonoBehaviour
             if (!bgMusic[music].isPlaying) {
                 StopMusic();
                 bgMusic[music].Play();
-            } else {
-                print("Music continues between scenes");
             }
         } else {
             Debug.Log("Error: int variable supplied lies out of soundtrack index");
