+++ Assets/Scripts/SceneAudio.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 22c6ad5aa2f9e4ceea2c2085ce0bd150
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
